+++ backend/src/main/java/de/bund/digitalservice/ris/caselaw/adapter/transformer/ldml/CommonPortalTransformer.java
@@ -1,4 +1,4 @@
-package de.bund.digitalservice.ris.caselaw.adapter.transformer;
+package de.bund.digitalservice.ris.caselaw.adapter.transformer.ldml;
 
 import static de.bund.digitalservice.ris.caselaw.adapter.MappingUtils.applyIfNotEmpty;
 import static de.bund.digitalservice.ris.caselaw.adapter.MappingUtils.nullSafeGet;
